Fix exception being thrown in AutoloadSourceLocator
The handling in FunctionReflector was actually wrong - it was returning `null`
when it should be throwing an `IdentifierNotFound` exception. In the case of
using AutoloadSourceLocator when reflecting on a function, an exception
`FunctionUndefined` was being thrown. This has been removed (therefore this is
sadly a BC break) and we handle this correctly.