Minor performance optimization: looking for constants in more recent includes first
Constants could be defined by any file, and we often want to look at constants defined
in own code/libraries rather than in frameworks/libraries that were used to bootstrap
and analyse our code (such as better-reflection itself).

In practice, this is a "guessed optimisation" that will likely affect the performance
of most consumers of the library, and it will certainly remove the need for traversing
autoloaded files for `roave/better-reflection` and `nikic/php-parser` in most setups,
leading to a massive net performance benefit.

A miss on a constant lookup will still have a massive performance penalty.