Rewrote #901 so that `ReflectionUnionType` is kept internally for union types
When an union type of `T|null` or `?T` is met, BetterReflection will (from now on)
keep a `ReflectionUnionType` internally, while exposing a `ReflectionNamedType`
with `ReflectionNamedType#allowsNull() === true` only at adapter level.

While this is a BC break, it leads to a much cleaner API around handling `null`
types, and inspecting types for type analysis.

Ref: https://github.com/Roave/BetterReflection/pull/902#discussion_r762529382
Ref: https://github.com/Roave/BackwardCompatibilityCheck/pull/324