Feature Request: Reflect All Attributes
A bare-minimum, proof-of-concept implementation for gathering all
ReflectionAttributes from a given source. This PR updates the Reflector
interface and DefaultReflector implementation to include a
reflectAllAttributes() method that will return a list of
ReflectionAttribute. Additionally, several type-hints on parameters and
method returns have been updated to include Node\Attribute and
ReflectionAttribute where appropriate.

This implementation is not meant to represent something fully working or
ready to be merged in but to spark a conversation on the desired
feature.