Remove unnecessary constants
These constants came from the parent classes, and they're exactly the same as the parent ones.
The issue is they conflict on PHP 8.4, because of the "int" type is being added there.

Since the values are exactly the same, and there is no usage different; This will remove the unneeded constants.